Startpage

+++ src/components/games.js
@@ -1,12 +1,13 @@
-import React, { createElement } from 'react';
+import React from 'react';
 
 
 class Games extends React.Component {
     
     constructor(){
     super();
-    this.state = {notice: false };
+    this.state = {notice: []};
     this.checkGames= this.checkGames.bind(this);
+    this.showPlayer= this.showPlayer.bind(this);
     this.addGame = this.addGame.bind(this);
     this.joinGame = this.joinGame.bind(this);
     this.deleteGame= this.deleteGame.bind(this);
@@ -30,9 +31,9 @@ class Games extends React.Component {
                 for(var i = 0; i < data.games.length; i++){
                     if (data.games[i].running === false){
 
-                        console.log("joining available game " + JSON.stringify(data.games[i].id))
+                        console.log("joining available game " + JSON.stringify(data.games[i]))
                         //joinGame(0/*wird wieder vom spieler übernommen*/);
-                        this.setState({notice: data});
+                        // this.setState({notice: data});
                     }
                 }
                 
@@ -45,7 +46,36 @@ class Games extends React.Component {
 
     componentDidMount() {
        this.checkGames();
-     }
+       this.showPlayer();
+    }
+    
+    showPlayer(){
+
+        const serviceendpoint = "https://gruppe7.toni-barth.com";
+
+        fetch(serviceendpoint + '/games/')
+        .then(response => response.json())
+        .then(data=>{
+            console.log('Im Here!');
+            if(data.games.length !== 0)
+            {
+                for(var i = 0; i < data.games.length; i++){
+                    if (data.games[i].running === false){
+                        console.log('PLAYER IN GAME(' + data.games[i].id + '): '+ JSON.stringify(data.games[i].players));
+                        this.setState({notice: data.games[i].players});
+                        console.log(data.games[i].players.length + ' Player/s in Game.');
+                    }
+                }
+            }
+        })
+        .catch((error) => {
+            console.error('Error:', error);
+            });
+    } 
+
+    componentDidUpdate() {
+       this.showPlayer();
+    }
 
     addGame() {
 
@@ -54,27 +84,30 @@ class Games extends React.Component {
         fetch(serviceendpoint + '/players/')
         .then(response => response.json())
         .then(data=>{
-           var id= data.players[data.players.length-1].id;
-            console.log("PlayerID:" + id);
+            if(data.players.length !== 0){
+                var id= data.players[data.players.length-1].id;
+                console.log("PlayerID:" + id);
 
-            //neues Spiel wird erstellt mit eigener SpielerID
-            fetch(serviceendpoint + '/games/', {
-                method: "POST",
-                body: JSON.stringify({ owner: id}),
-                headers: { "Content-Type": "application/json" }
-            })
-            .then(response => response.json())
-            .then(data=>{
-                console.log(data.owner.name)
-            })
-            .catch((error) => {
-                console.error('Error:', error);
-            });
+                //neues Spiel wird erstellt mit eigener SpielerID
+                fetch(serviceendpoint + '/games/', {
+                    method: "POST",
+                    body: JSON.stringify({ owner: id}),
+                    headers: { "Content-Type": "application/json" }
+                })
+                .then(response => response.json())
+                .then(data=>{
+                    console.log(data.owner.name)
+                })
+                .catch((error) => {
+                    console.error('Error:', error);
+                });
+            }else{
+                console.log('No Player existing.');
+            }   
         })
         .catch((error) => {
             console.error('Error:', error);
-            });
-
+        });
         
     }
 
@@ -104,30 +137,39 @@ class Games extends React.Component {
 
     deleteGame(){
 
-
         const serviceendpoint = "https://gruppe7.toni-barth.com";
         fetch(serviceendpoint + '/games/')
         .then(response => response.json())
-        .then(data => {var gameID = data.games[data.games.length-1].id;
-        
-        fetch(serviceendpoint + '/games/' + gameID, {
-            method: "DELETE",
-            headers: { "Content-Type": "application/json" }
-        })
-        .then(response => response.json())
-        .catch((error) => {
-            console.error('Error:', error);
-            });
+        .then(data => {
+            if(data.games.length !== 0){
+                var gameID = data.games[data.games.length-1].id;
+            
+                fetch(serviceendpoint + '/games/' + gameID, {
+                    method: "DELETE",
+                    headers: { "Content-Type": "application/json" }
+                })
+                .then(response => response.json())
+                .catch((error) => {
+                    console.error('Error:', error);
+                });
+            }else{
+                console.log('No Games existing.');
+            }    
         })
 
     }
 
     render() {
         return (
-            <div>
-                <button className='button' id= "newGameBtn" onClick={this.addGame}>New Game</button>
-                <button className='button' id= "delGameBtn" onClick={this.deleteGame}>Delete Game</button>
-            </div>
+            <>
+                <button id= "newGameBtn" className='newBtn' onClick={this.addGame}>New Game</button>
+                <button id= "delGameBtn" className='delBtn' onClick={this.deleteGame}>Delete Game</button>
+                <div className='list'>
+                    {this.state.notice.map(({ name, id }) => (
+                        <p key={id}> {name}</p>
+                    ))}
+                </div>
+            </>
         );
     }
 }
